+++ src/main/java/com/aitusoftware/network/patterns/measurement/LatencyRecorder.java
@@ -5,5 +5,5 @@ public interface LatencyRecorder
     void recordValue(final long value);
     void reset();
     void complete();
-    void messagesDropped(final int messageCount);
+    void messagesDropped(final long messageCount);
 }
